Fixed Inquiry Styles, Refactored css files to styles folder, got rid of the alert when adding to cart from menu and menuitem
